+++ Flower_shop/EfStuff/Repositories/Interfaces/IBaseRepository.cs
@@ -4,9 +4,9 @@
     {
         Task<bool> AnyAsync();
 
-        T GetById(int id);
+        T GetById(long id);
 
-        Task<T> GetByIdAsync(int id);
+        Task<T> GetByIdAsync(long id);
 
         Task<List<T>> GetAllAsync();
 
@@ -16,7 +16,7 @@
 
         Task RemoveAsync(T model);
 
-        Task RemoveByIdAsync(int id);
+        Task RemoveByIdAsync(long id);
 
         Task<int> CountAsync();
     }
